requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
playwright>=1.40.0
pyyaml>=6.0
jinja2>=3.1.0
weasyprint>=60.0
matplotlib>=3.8.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.6
notion-client>=2.0.0
